move conversion functionality to dense
