+++ core/device_hooks/common_kernels.inc.cpp
@@ -457,6 +457,20 @@ namespace dense {
 
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_DENSE_SIMPLE_APPLY_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_DENSE_APPLY_KERNEL);
+GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_DENSE_CONVERT_TO_COO_KERNEL);
+GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_DENSE_CONVERT_TO_CSR_KERNEL);
+GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_DENSE_CONVERT_TO_ELL_KERNEL);
+GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_DENSE_CONVERT_TO_FBCSR_KERNEL);
+GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_DENSE_CONVERT_TO_HYBRID_KERNEL);
+GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_DENSE_CONVERT_TO_SELLP_KERNEL);
+GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_DENSE_CONVERT_TO_SPARSITY_CSR_KERNEL);
+GKO_STUB_VALUE_TYPE(GKO_DECLARE_DENSE_COMPUTE_MAX_NNZ_PER_ROW_KERNEL);
+GKO_STUB_VALUE_TYPE(GKO_DECLARE_DENSE_COMPUTE_SLICE_SETS_KERNEL);
+GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_DENSE_COUNT_NONZEROS_PER_ROW_KERNEL);
+GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
+    GKO_DECLARE_DENSE_COUNT_NONZEROS_PER_ROW_KERNEL_SIZE_T);
+GKO_STUB_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_COUNT_NONZERO_BLOCKS_PER_ROW_KERNEL);
 
 
 }  // namespace dense
@@ -487,22 +501,6 @@ GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_COMPUTE_SQUARED_NORM2_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_COMPUTE_SQRT_KERNEL);
 GKO_STUB_VALUE_AND_INDEX_TYPE(
     GKO_DECLARE_MULTIVECTOR_FILL_IN_MATRIX_DATA_KERNEL);
-GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_MULTIVECTOR_CONVERT_TO_COO_KERNEL);
-GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_MULTIVECTOR_CONVERT_TO_CSR_KERNEL);
-GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_MULTIVECTOR_CONVERT_TO_ELL_KERNEL);
-GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_MULTIVECTOR_CONVERT_TO_FBCSR_KERNEL);
-GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_MULTIVECTOR_CONVERT_TO_HYBRID_KERNEL);
-GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_MULTIVECTOR_CONVERT_TO_SELLP_KERNEL);
-GKO_STUB_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_CONVERT_TO_SPARSITY_CSR_KERNEL);
-GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_COMPUTE_MAX_NNZ_PER_ROW_KERNEL);
-GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_COMPUTE_SLICE_SETS_KERNEL);
-GKO_STUB_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_COUNT_NONZEROS_PER_ROW_KERNEL);
-GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
-    GKO_DECLARE_MULTIVECTOR_COUNT_NONZEROS_PER_ROW_KERNEL_SIZE_T);
-GKO_STUB_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_COUNT_NONZERO_BLOCKS_PER_ROW_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_TRANSPOSE_KERNEL);
 GKO_STUB_VALUE_TYPE(GKO_DECLARE_MULTIVECTOR_CONJ_TRANSPOSE_KERNEL);
 GKO_STUB_VALUE_AND_INDEX_TYPE(GKO_DECLARE_MULTIVECTOR_SYMM_PERMUTE_KERNEL);
