+++ core/matrix/coo.cpp
@@ -14,6 +14,7 @@
 #include <ginkgo/core/base/temporary_clone.hpp>
 #include <ginkgo/core/base/utils.hpp>
 #include <ginkgo/core/matrix/csr.hpp>
+#include <ginkgo/core/matrix/dense.hpp>
 #include <ginkgo/core/matrix/multivector.hpp>
 
 #include "core/base/device_matrix_data_kernels.hpp"
@@ -331,7 +332,7 @@ void Coo<ValueType, IndexType>::move_to(Csr<ValueType, IndexType>* result)
 
 
 template <typename ValueType, typename IndexType>
-void Coo<ValueType, IndexType>::convert_to(MultiVector<ValueType>* result) const
+void Coo<ValueType, IndexType>::convert_to(Dense<ValueType>* result) const
 {
     auto exec = this->get_executor();
     auto tmp_result = make_temporary_output_clone(exec, result);
@@ -343,7 +344,7 @@ void Coo<ValueType, IndexType>::convert_to(MultiVector<ValueType>* result) const
 
 
 template <typename ValueType, typename IndexType>
-void Coo<ValueType, IndexType>::move_to(MultiVector<ValueType>* result)
+void Coo<ValueType, IndexType>::move_to(Dense<ValueType>* result)
 {
     this->convert_to(result);
 }
