+++ core/matrix/ell.cpp
@@ -13,6 +13,7 @@
 #include <ginkgo/core/base/temporary_clone.hpp>
 #include <ginkgo/core/base/utils.hpp>
 #include <ginkgo/core/matrix/csr.hpp>
+#include <ginkgo/core/matrix/dense.hpp>
 #include <ginkgo/core/matrix/device_views.hpp>
 #include <ginkgo/core/matrix/multivector.hpp>
 
@@ -297,7 +298,7 @@ void Ell<ValueType, IndexType>::move_to(
 
 
 template <typename ValueType, typename IndexType>
-void Ell<ValueType, IndexType>::convert_to(MultiVector<ValueType>* result) const
+void Ell<ValueType, IndexType>::convert_to(Dense<ValueType>* result) const
 {
     auto exec = this->get_executor();
     auto tmp_result = make_temporary_output_clone(exec, result);
@@ -309,7 +310,7 @@ void Ell<ValueType, IndexType>::convert_to(MultiVector<ValueType>* result) const
 
 
 template <typename ValueType, typename IndexType>
-void Ell<ValueType, IndexType>::move_to(MultiVector<ValueType>* result)
+void Ell<ValueType, IndexType>::move_to(Dense<ValueType>* result)
 {
     this->convert_to(result);
 }
