+++ core/matrix/hybrid.cpp
@@ -12,6 +12,7 @@
 #include <ginkgo/core/base/precision_dispatch.hpp>
 #include <ginkgo/core/base/temporary_clone.hpp>
 #include <ginkgo/core/base/utils.hpp>
+#include <ginkgo/core/matrix/dense.hpp>
 #include <ginkgo/core/matrix/multivector.hpp>
 
 #include "core/base/array_access.hpp"
@@ -286,8 +287,7 @@ void Hybrid<ValueType, IndexType>::move_to(
 
 
 template <typename ValueType, typename IndexType>
-void Hybrid<ValueType, IndexType>::convert_to(
-    MultiVector<ValueType>* result) const
+void Hybrid<ValueType, IndexType>::convert_to(Dense<ValueType>* result) const
 {
     auto exec = this->get_executor();
     result->resize(this->get_size());
@@ -303,7 +303,7 @@ void Hybrid<ValueType, IndexType>::convert_to(
 
 
 template <typename ValueType, typename IndexType>
-void Hybrid<ValueType, IndexType>::move_to(MultiVector<ValueType>* result)
+void Hybrid<ValueType, IndexType>::move_to(Dense<ValueType>* result)
 {
     this->convert_to(result);
 }
