+++ core/matrix/multivector.cpp
@@ -15,23 +15,13 @@
 #include <ginkgo/core/base/precision_dispatch.hpp>
 #include <ginkgo/core/base/temporary_clone.hpp>
 #include <ginkgo/core/base/utils.hpp>
-#include <ginkgo/core/matrix/coo.hpp>
-#include <ginkgo/core/matrix/csr.hpp>
 #include <ginkgo/core/matrix/dense.hpp>
 #include <ginkgo/core/matrix/diagonal.hpp>
-#include <ginkgo/core/matrix/ell.hpp>
-#include <ginkgo/core/matrix/fbcsr.hpp>
-#include <ginkgo/core/matrix/hybrid.hpp>
 #include <ginkgo/core/matrix/permutation.hpp>
 #include <ginkgo/core/matrix/scaled_permutation.hpp>
-#include <ginkgo/core/matrix/sellp.hpp>
-#include <ginkgo/core/matrix/sparsity_csr.hpp>
 
-#include "core/base/array_access.hpp"
 #include "core/base/dispatch_helper.hpp"
 #include "core/base/validation.hpp"
-#include "core/components/prefix_sum_kernels.hpp"
-#include "core/matrix/hybrid_kernels.hpp"
 #include "core/matrix/multivector_kernels.hpp"
 #include "core/matrix/permutation.hpp"
 
@@ -59,17 +49,6 @@ GKO_REGISTER_OPERATION(compute_mean, multivector::compute_mean);
 GKO_REGISTER_OPERATION(compute_squared_norm2,
                        multivector::compute_squared_norm2);
 GKO_REGISTER_OPERATION(compute_sqrt, multivector::compute_sqrt);
-GKO_REGISTER_OPERATION(compute_max_nnz_per_row,
-                       multivector::compute_max_nnz_per_row);
-GKO_REGISTER_OPERATION(compute_hybrid_coo_row_ptrs,
-                       hybrid::compute_coo_row_ptrs);
-GKO_REGISTER_OPERATION(count_nonzeros_per_row,
-                       multivector::count_nonzeros_per_row);
-GKO_REGISTER_OPERATION(count_nonzero_blocks_per_row,
-                       multivector::count_nonzero_blocks_per_row);
-GKO_REGISTER_OPERATION(prefix_sum_nonnegative,
-                       components::prefix_sum_nonnegative);
-GKO_REGISTER_OPERATION(compute_slice_sets, multivector::compute_slice_sets);
 GKO_REGISTER_OPERATION(transpose, multivector::transpose);
 GKO_REGISTER_OPERATION(conj_transpose, multivector::conj_transpose);
 GKO_REGISTER_OPERATION(symm_permute, multivector::symm_permute);
@@ -95,14 +74,6 @@ GKO_REGISTER_OPERATION(inv_row_scale_permute,
 GKO_REGISTER_OPERATION(inv_col_scale_permute,
                        multivector::inv_col_scale_permute);
 GKO_REGISTER_OPERATION(fill_in_matrix_data, multivector::fill_in_matrix_data);
-GKO_REGISTER_OPERATION(convert_to_coo, multivector::convert_to_coo);
-GKO_REGISTER_OPERATION(convert_to_csr, multivector::convert_to_csr);
-GKO_REGISTER_OPERATION(convert_to_ell, multivector::convert_to_ell);
-GKO_REGISTER_OPERATION(convert_to_fbcsr, multivector::convert_to_fbcsr);
-GKO_REGISTER_OPERATION(convert_to_hybrid, multivector::convert_to_hybrid);
-GKO_REGISTER_OPERATION(convert_to_sellp, multivector::convert_to_sellp);
-GKO_REGISTER_OPERATION(convert_to_sparsity_csr,
-                       multivector::convert_to_sparsity_csr);
 GKO_REGISTER_OPERATION(extract_diagonal, multivector::extract_diagonal);
 GKO_REGISTER_OPERATION(inplace_absolute_dense,
                        multivector::inplace_absolute_dense);
@@ -746,109 +717,6 @@ void MultiVector<ValueType>::move_to(
 
 
 template <typename ValueType>
-template <typename IndexType>
-void MultiVector<ValueType>::convert_impl(
-    Coo<ValueType, IndexType>* result) const
-{
-    auto exec = this->get_executor();
-    const auto num_rows = this->get_size()[0];
-
-    array<int64> row_ptrs{exec, num_rows + 1};
-    exec->run(multivector::make_count_nonzeros_per_row(
-        this->get_const_device_view(), row_ptrs.get_data()));
-    exec->run(multivector::make_prefix_sum_nonnegative(row_ptrs.get_data(),
-                                                       num_rows + 1));
-    const auto nnz = get_element(row_ptrs, num_rows);
-    result->resize(this->get_size(), nnz);
-    exec->run(multivector::make_convert_to_coo(
-        this->get_const_device_view(), row_ptrs.get_const_data(),
-        make_temporary_clone(exec, result)->get_device_view()));
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Coo<ValueType, int32>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Coo<ValueType, int32>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Coo<ValueType, int64>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Coo<ValueType, int64>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-template <typename IndexType>
-void MultiVector<ValueType>::convert_impl(
-    Csr<ValueType, IndexType>* result) const
-{
-    {
-        auto exec = this->get_executor();
-        const auto num_rows = this->get_size()[0];
-        auto tmp = make_temporary_clone(exec, result);
-        tmp->row_ptrs_.resize_and_reset(num_rows + 1);
-        exec->run(multivector::make_count_nonzeros_per_row(
-            this->get_const_device_view(), tmp->get_row_ptrs()));
-        exec->run(multivector::make_prefix_sum_nonnegative(tmp->get_row_ptrs(),
-                                                           num_rows + 1));
-        const auto nnz =
-            exec->copy_val_to_host(tmp->get_const_row_ptrs() + num_rows);
-        tmp->col_idxs_.resize_and_reset(nnz);
-        tmp->values_.resize_and_reset(nnz);
-        tmp->set_size(this->get_size());
-        exec->run(multivector::make_convert_to_csr(
-            this->get_const_device_view(), tmp->get_device_view()));
-    }
-    result->make_srow();
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Csr<ValueType, int32>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Csr<ValueType, int32>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Csr<ValueType, int64>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Csr<ValueType, int64>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
 void MultiVector<ValueType>::convert_to(Dense<ValueType>* result) const
 {
     if (result->get_size() != this->get_size()) {
@@ -871,270 +739,6 @@ void MultiVector<ValueType>::move_to(Dense<ValueType>* result)
     this->set_size(dim<2>{0, 0});
     result->stride_ = std::exchange(stride_, 0);
     result->values_ = std::move(values_);
-}
-
-
-template <typename ValueType>
-template <typename IndexType>
-void MultiVector<ValueType>::convert_impl(
-    Fbcsr<ValueType, IndexType>* result) const
-{
-    auto exec = this->get_executor();
-    const auto bs = result->get_block_size();
-    const auto row_blocks = detail::get_num_blocks(bs, this->get_size()[0]);
-    const auto col_blocks = detail::get_num_blocks(bs, this->get_size()[1]);
-    auto tmp = make_temporary_clone(exec, result);
-    tmp->row_ptrs_.resize_and_reset(row_blocks + 1);
-    exec->run(multivector::make_count_nonzero_blocks_per_row(
-        this->get_const_device_view(), bs, tmp->get_row_ptrs()));
-    exec->run(multivector::make_prefix_sum_nonnegative(tmp->get_row_ptrs(),
-                                                       row_blocks + 1));
-    const auto nnz_blocks =
-        exec->copy_val_to_host(tmp->get_const_row_ptrs() + row_blocks);
-    tmp->col_idxs_.resize_and_reset(nnz_blocks);
-    tmp->values_.resize_and_reset(nnz_blocks * bs * bs);
-    tmp->values_.fill(zero<ValueType>());
-    tmp->set_size(this->get_size());
-    exec->run(multivector::make_convert_to_fbcsr(this->get_const_device_view(),
-                                                 tmp.get()));
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Fbcsr<ValueType, int32>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Fbcsr<ValueType, int32>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Fbcsr<ValueType, int64>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Fbcsr<ValueType, int64>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-template <typename IndexType>
-void MultiVector<ValueType>::convert_impl(
-    Ell<ValueType, IndexType>* result) const
-{
-    auto exec = this->get_executor();
-    size_type num_stored_elements_per_row{};
-    exec->run(multivector::make_compute_max_nnz_per_row(
-        this->get_const_device_view(), num_stored_elements_per_row));
-    result->resize(this->get_size(), num_stored_elements_per_row);
-    exec->run(multivector::make_convert_to_ell(
-        this->get_const_device_view(),
-        make_temporary_clone(exec, result)->get_device_view()));
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Ell<ValueType, int32>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Ell<ValueType, int32>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Ell<ValueType, int64>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Ell<ValueType, int64>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-template <typename IndexType>
-void MultiVector<ValueType>::convert_impl(
-    Hybrid<ValueType, IndexType>* result) const
-{
-    auto exec = this->get_executor();
-    const auto num_rows = this->get_size()[0];
-    const auto num_cols = this->get_size()[1];
-    array<size_type> row_nnz{exec, num_rows};
-    array<int64> coo_row_ptrs{exec, num_rows + 1};
-    exec->run(multivector::make_count_nonzeros_per_row(
-        this->get_const_device_view(), row_nnz.get_data()));
-    size_type ell_lim{};
-    size_type coo_nnz{};
-    result->get_strategy()->compute_hybrid_config(row_nnz, &ell_lim, &coo_nnz);
-    if (ell_lim > num_cols) {
-        // TODO remove temporary fix after ELL gains true structural zeros
-        ell_lim = num_cols;
-    }
-    exec->run(multivector::make_compute_hybrid_coo_row_ptrs(
-        row_nnz, ell_lim, coo_row_ptrs.get_data()));
-    coo_nnz = get_element(coo_row_ptrs, num_rows);
-    auto tmp = make_temporary_clone(exec, result);
-    tmp->resize(this->get_size(), ell_lim, coo_nnz);
-    exec->run(multivector::make_convert_to_hybrid(this->get_const_device_view(),
-                                                  coo_row_ptrs.get_const_data(),
-                                                  tmp->get_device_view()));
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Hybrid<ValueType, int32>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Hybrid<ValueType, int32>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Hybrid<ValueType, int64>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Hybrid<ValueType, int64>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-template <typename IndexType>
-void MultiVector<ValueType>::convert_impl(
-    Sellp<ValueType, IndexType>* result) const
-{
-    auto exec = this->get_executor();
-    const auto num_rows = this->get_size()[0];
-    const auto stride_factor = result->get_stride_factor();
-    const auto slice_size = result->get_slice_size();
-    const auto num_slices = ceildiv(num_rows, slice_size);
-    auto tmp = make_temporary_clone(exec, result);
-    tmp->stride_factor_ = stride_factor;
-    tmp->slice_size_ = slice_size;
-    tmp->slice_sets_.resize_and_reset(num_slices + 1);
-    tmp->slice_lengths_.resize_and_reset(num_slices);
-    exec->run(multivector::make_compute_slice_sets(
-        this->get_const_device_view(), slice_size, stride_factor,
-        tmp->get_slice_sets(), tmp->get_slice_lengths()));
-    auto total_cols =
-        exec->copy_val_to_host(tmp->get_slice_sets() + num_slices);
-    tmp->col_idxs_.resize_and_reset(total_cols * slice_size);
-    tmp->values_.resize_and_reset(total_cols * slice_size);
-    tmp->set_size(this->get_size());
-    exec->run(multivector::make_convert_to_sellp(this->get_const_device_view(),
-                                                 tmp->get_device_view()));
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Sellp<ValueType, int32>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Sellp<ValueType, int32>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(Sellp<ValueType, int64>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(Sellp<ValueType, int64>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-template <typename IndexType>
-void MultiVector<ValueType>::convert_impl(
-    SparsityCsr<ValueType, IndexType>* result) const
-{
-    auto exec = this->get_executor();
-    const auto num_rows = this->get_size()[0];
-    auto tmp = make_temporary_clone(exec, result);
-    tmp->row_ptrs_.resize_and_reset(num_rows + 1);
-    exec->run(multivector::make_count_nonzeros_per_row(
-        this->get_const_device_view(), tmp->row_ptrs_.get_data()));
-    exec->run(multivector::make_prefix_sum_nonnegative(
-        tmp->row_ptrs_.get_data(), num_rows + 1));
-    const auto nnz = get_element(tmp->row_ptrs_, num_rows);
-    tmp->col_idxs_.resize_and_reset(nnz);
-    tmp->value_.fill(one<ValueType>());
-    tmp->set_size(this->get_size());
-    exec->run(multivector::make_convert_to_sparsity_csr(
-        this->get_const_device_view(), tmp.get()));
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(
-    SparsityCsr<ValueType, int32>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(SparsityCsr<ValueType, int32>* result)
-{
-    this->convert_to(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::convert_to(
-    SparsityCsr<ValueType, int64>* result) const
-{
-    this->convert_impl(result);
-}
-
-
-template <typename ValueType>
-void MultiVector<ValueType>::move_to(SparsityCsr<ValueType, int64>* result)
-{
-    this->convert_to(result);
 }
