+++ core/matrix/sellp.cpp
@@ -10,6 +10,7 @@
 #include <ginkgo/core/base/precision_dispatch.hpp>
 #include <ginkgo/core/base/utils.hpp>
 #include <ginkgo/core/matrix/csr.hpp>
+#include <ginkgo/core/matrix/dense.hpp>
 #include <ginkgo/core/matrix/multivector.hpp>
 
 #include "core/base/allocator.hpp"
@@ -277,8 +278,7 @@ void Sellp<ValueType, IndexType>::move_to(
 
 
 template <typename ValueType, typename IndexType>
-void Sellp<ValueType, IndexType>::convert_to(
-    MultiVector<ValueType>* result) const
+void Sellp<ValueType, IndexType>::convert_to(Dense<ValueType>* result) const
 {
     auto exec = this->get_executor();
     auto tmp_result = make_temporary_output_clone(exec, result);
@@ -290,7 +290,7 @@ void Sellp<ValueType, IndexType>::convert_to(
 
 
 template <typename ValueType, typename IndexType>
-void Sellp<ValueType, IndexType>::move_to(MultiVector<ValueType>* result)
+void Sellp<ValueType, IndexType>::move_to(Dense<ValueType>* result)
 {
     this->convert_to(result);
 }
