+++ dpcpp/matrix/dense_kernels.dp.cpp
@@ -83,7 +83,7 @@ void apply(std::shared_ptr<const DefaultExecutor> exec,
                     exec->copy_val_to_host(beta.values),
                     as_device_type(c.values), c.stride);
             } else {
-                dense::scale(exec, beta, c);
+                multivector::scale(exec, beta, c);
             }
         }
     } else {
@@ -92,6 +92,266 @@ void apply(std::shared_ptr<const DefaultExecutor> exec,
 }
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(GKO_DECLARE_DENSE_APPLY_KERNEL);
+
+
+template <typename ValueType, typename IndexType>
+void convert_to_coo(std::shared_ptr<const DefaultExecutor> exec,
+                    matrix::view::dense<const ValueType> source,
+                    const int64* row_ptrs,
+                    matrix::view::coo<ValueType, IndexType> result)
+{
+    const auto num_rows = result.size[0];
+    const auto num_cols = result.size[1];
+    const auto in_vals = as_device_type(source.values);
+    const auto stride = source.stride;
+
+    auto rows = result.row_idxs;
+    auto cols = result.col_idxs;
+    auto vals = as_device_type(result.values);
+
+    exec->get_queue()->submit([&](sycl::handler& cgh) {
+        cgh.parallel_for(num_rows, [=](sycl::item<1> item) {
+            const auto row = static_cast<size_type>(item[0]);
+            auto write_to = row_ptrs[row];
+
+            for (size_type col = 0; col < num_cols; col++) {
+                if (is_nonzero(in_vals[stride * row + col])) {
+                    vals[write_to] = in_vals[stride * row + col];
+                    cols[write_to] = static_cast<IndexType>(col);
+                    rows[write_to] = static_cast<IndexType>(row);
+                    write_to++;
+                }
+            }
+        });
+    });
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_CONVERT_TO_COO_KERNEL);
+
+
+template <typename ValueType, typename IndexType>
+void convert_to_csr(std::shared_ptr<const DefaultExecutor> exec,
+                    matrix::view::dense<const ValueType> source,
+                    matrix::view::csr<ValueType, IndexType> result)
+{
+    const auto num_rows = result.size[0];
+    const auto num_cols = result.size[1];
+    const auto in_vals = as_device_type(source.values);
+    const auto stride = source.stride;
+
+    const auto row_ptrs = result.row_ptrs;
+    auto cols = result.col_idxs;
+    auto vals = as_device_type(result.values);
+
+    exec->get_queue()->submit([&](sycl::handler& cgh) {
+        cgh.parallel_for(num_rows, [=](sycl::item<1> item) {
+            const auto row = static_cast<size_type>(item[0]);
+            auto write_to = row_ptrs[row];
+
+            for (size_type col = 0; col < num_cols; col++) {
+                if (is_nonzero(in_vals[stride * row + col])) {
+                    vals[write_to] = in_vals[stride * row + col];
+                    cols[write_to] = static_cast<IndexType>(col);
+                    write_to++;
+                }
+            }
+        });
+    });
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_CONVERT_TO_CSR_KERNEL);
+
+
+template <typename ValueType, typename IndexType>
+void convert_to_ell(std::shared_ptr<const DefaultExecutor> exec,
+                    matrix::view::dense<const ValueType> source,
+                    matrix::view::ell<ValueType, IndexType> result)
+{
+    const auto num_rows = result.size[0];
+    const auto num_cols = result.size[1];
+    const auto max_nnz_per_row = result.num_stored_elements_per_row;
+    const auto in_vals = as_device_type(source.values);
+    const auto in_stride = source.stride;
+
+    auto cols = result.col_idxs;
+    auto vals = as_device_type(result.values);
+    const auto stride = result.stride;
+
+    exec->get_queue()->submit([&](sycl::handler& cgh) {
+        cgh.parallel_for(num_rows, [=](sycl::item<1> item) {
+            const auto row = static_cast<size_type>(item[0]);
+            size_type col_idx = 0;
+            for (size_type col = 0; col < num_cols; col++) {
+                if (is_nonzero(in_vals[row * in_stride + col])) {
+                    cols[col_idx * stride + row] = col;
+                    vals[col_idx * stride + row] =
+                        in_vals[row * in_stride + col];
+                    col_idx++;
+                }
+            }
+            for (; col_idx < max_nnz_per_row; col_idx++) {
+                cols[col_idx * stride + row] = invalid_index<IndexType>();
+                vals[col_idx * stride + row] = zero<device_type<ValueType>>();
+            }
+        });
+    });
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_CONVERT_TO_ELL_KERNEL);
+
+
+template <typename ValueType, typename IndexType>
+void convert_to_fbcsr(std::shared_ptr<const DefaultExecutor> exec,
+                      matrix::view::dense<const ValueType> source,
+                      matrix::Fbcsr<ValueType, IndexType>* result)
+    GKO_NOT_IMPLEMENTED;
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_CONVERT_TO_FBCSR_KERNEL);
+
+
+template <typename ValueType, typename IndexType>
+void count_nonzero_blocks_per_row(std::shared_ptr<const DefaultExecutor> exec,
+                                  matrix::view::dense<const ValueType> source,
+                                  int bs,
+                                  IndexType* result) GKO_NOT_IMPLEMENTED;
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_COUNT_NONZERO_BLOCKS_PER_ROW_KERNEL);
+
+
+template <typename ValueType, typename IndexType>
+void convert_to_hybrid(std::shared_ptr<const DefaultExecutor> exec,
+                       matrix::view::dense<const ValueType> source,
+                       const int64* coo_row_ptrs,
+                       matrix::view::hybrid<ValueType, IndexType> result)
+{
+    const auto num_rows = result.size[0];
+    const auto num_cols = result.size[1];
+    const auto ell_lim = result.ell_part.num_stored_elements_per_row;
+    const auto in_vals = as_device_type(source.values);
+    const auto in_stride = source.stride;
+    const auto ell_stride = result.ell_part.stride;
+    auto ell_cols = result.ell_part.col_idxs;
+    auto ell_vals = as_device_type(result.ell_part.values);
+    auto coo_rows = result.coo_part.row_idxs;
+    auto coo_cols = result.coo_part.col_idxs;
+    auto coo_vals = as_device_type(result.coo_part.values);
+
+    exec->get_queue()->submit([&](sycl::handler& cgh) {
+        cgh.parallel_for(num_rows, [=](sycl::item<1> item) {
+            const auto row = static_cast<size_type>(item[0]);
+            size_type ell_count = 0;
+            size_type col = 0;
+            auto ell_idx = row;
+            for (; col < num_cols && ell_count < ell_lim; col++) {
+                const auto val = in_vals[row * in_stride + col];
+                if (is_nonzero(val)) {
+                    ell_vals[ell_idx] = val;
+                    ell_cols[ell_idx] = static_cast<IndexType>(col);
+                    ell_count++;
+                    ell_idx += ell_stride;
+                }
+            }
+            for (; ell_count < ell_lim; ell_count++) {
+                ell_vals[ell_idx] = zero<device_type<ValueType>>();
+                ell_cols[ell_idx] = invalid_index<IndexType>();
+                ell_idx += ell_stride;
+            }
+            auto coo_idx = coo_row_ptrs[row];
+            for (; col < num_cols; col++) {
+                const auto val = in_vals[row * in_stride + col];
+                if (is_nonzero(val)) {
+                    coo_vals[coo_idx] = val;
+                    coo_cols[coo_idx] = static_cast<IndexType>(col);
+                    coo_rows[coo_idx] = static_cast<IndexType>(row);
+                    coo_idx++;
+                }
+            }
+        });
+    });
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_CONVERT_TO_HYBRID_KERNEL);
+
+
+template <typename ValueType, typename IndexType>
+void convert_to_sellp(std::shared_ptr<const DefaultExecutor> exec,
+                      matrix::view::dense<const ValueType> source,
+                      matrix::view::sellp<ValueType, IndexType> result)
+{
+    const auto num_rows = result.size[0];
+    const auto num_cols = result.size[1];
+    const auto stride = source.stride;
+    const auto in_vals = as_device_type(source.values);
+
+    const auto slice_sets = result.slice_sets;
+    const auto slice_size = result.slice_size;
+    auto vals = as_device_type(result.values);
+    auto col_idxs = result.col_idxs;
+
+    exec->get_queue()->submit([&](sycl::handler& cgh) {
+        cgh.parallel_for(num_rows, [=](sycl::item<1> item) {
+            const auto row = static_cast<size_type>(item[0]);
+            const auto local_row = row % slice_size;
+            const auto slice = row / slice_size;
+            const auto slice_end = slice_sets[slice + 1] * slice_size;
+            auto out_idx = slice_sets[slice] * slice_size + local_row;
+
+            for (size_type col = 0; col < num_cols; col++) {
+                const auto val = in_vals[row * stride + col];
+                if (is_nonzero(val)) {
+                    col_idxs[out_idx] = static_cast<IndexType>(col);
+                    vals[out_idx] = val;
+                    out_idx += slice_size;
+                }
+            }
+            for (; out_idx < slice_end; out_idx += slice_size) {
+                col_idxs[out_idx] = invalid_index<IndexType>();
+                vals[out_idx] = zero<device_type<ValueType>>();
+            }
+        });
+    });
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_CONVERT_TO_SELLP_KERNEL);
+
+
+template <typename ValueType, typename IndexType>
+void convert_to_sparsity_csr(std::shared_ptr<const DefaultExecutor> exec,
+                             matrix::view::dense<const ValueType> source,
+                             matrix::SparsityCsr<ValueType, IndexType>* result)
+{
+    const auto num_rows = result->get_size()[0];
+    const auto num_cols = result->get_size()[1];
+    const auto in_vals = as_device_type(source.values);
+    const auto stride = source.stride;
+
+    const auto row_ptrs = result->get_const_row_ptrs();
+    auto cols = result->get_col_idxs();
+
+    exec->get_queue()->submit([&](sycl::handler& cgh) {
+        cgh.parallel_for(num_rows, [=](sycl::item<1> item) {
+            const auto row = static_cast<size_type>(item[0]);
+            auto write_to = row_ptrs[row];
+
+            for (size_type col = 0; col < num_cols; col++) {
+                if (is_nonzero(in_vals[stride * row + col])) {
+                    cols[write_to] = static_cast<IndexType>(col);
+                    write_to++;
+                }
+            }
+        });
+    });
+}
+
+GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
+    GKO_DECLARE_DENSE_CONVERT_TO_SPARSITY_CSR_KERNEL);
 
 
 }  // namespace dense
