+++ include/ginkgo/core/matrix/hybrid.hpp
@@ -20,7 +20,7 @@ namespace matrix {
 
 
 template <typename ValueType>
-class MultiVector;
+class Dense;
 
 template <typename ValueType, typename IndexType>
 class Csr;
@@ -49,7 +49,7 @@ class Hybrid
 #if GINKGO_ENABLE_HALF && GINKGO_ENABLE_BFLOAT16
       public ConvertibleTo<Hybrid<next_precision<ValueType, 3>, IndexType>>,
 #endif
-      public ConvertibleTo<MultiVector<ValueType>>,
+      public ConvertibleTo<Dense<ValueType>>,
       public ConvertibleTo<Csr<ValueType, IndexType>>,
       public DiagonalExtractable<ValueType>,
       public ReadableFromMatrixData<ValueType, IndexType>,
@@ -57,7 +57,7 @@ class Hybrid
       public EnableAbsoluteComputation<
           remove_complex<Hybrid<ValueType, IndexType>>> {
     friend class EnableCloneable<Hybrid>;
-    friend class MultiVector<ValueType>;
+    friend class Dense<ValueType>;
     friend class Csr<ValueType, IndexType>;
     friend class Hybrid<to_complex<ValueType>, IndexType>;
     GKO_ASSERT_SUPPORTED_VALUE_AND_INDEX_TYPE;
@@ -68,8 +68,8 @@ public:
     using ConvertibleTo<
         Hybrid<next_precision<ValueType>, IndexType>>::convert_to;
     using ConvertibleTo<Hybrid<next_precision<ValueType>, IndexType>>::move_to;
-    using ConvertibleTo<MultiVector<ValueType>>::convert_to;
-    using ConvertibleTo<MultiVector<ValueType>>::move_to;
+    using ConvertibleTo<Dense<ValueType>>::convert_to;
+    using ConvertibleTo<Dense<ValueType>>::move_to;
     using ConvertibleTo<Csr<ValueType, IndexType>>::convert_to;
     using ConvertibleTo<Csr<ValueType, IndexType>>::move_to;
     using ReadableFromMatrixData<ValueType, IndexType>::read;
@@ -399,9 +399,9 @@ public:
         Hybrid<next_precision<ValueType, 3>, IndexType>* result) override;
 #endif
 
-    void convert_to(MultiVector<ValueType>* other) const override;
+    void convert_to(Dense<ValueType>* other) const override;
 
-    void move_to(MultiVector<ValueType>* other) override;
+    void move_to(Dense<ValueType>* other) override;
 
     void convert_to(Csr<ValueType, IndexType>* other) const override;
