+++ include/ginkgo/core/matrix/multivector.hpp
@@ -45,32 +45,9 @@ class VectorCache;
 namespace matrix {
 
 
-template <typename ValueType, typename IndexType>
-class Coo;
-
-template <typename ValueType, typename IndexType>
-class Csr;
-
-template <typename ValueType>
-class Diagonal;
-
-template <typename ValueType, typename IndexType>
-class Ell;
-
-template <typename ValueType, typename IndexType>
-class Fbcsr;
-
-template <typename ValueType, typename IndexType>
-class Hybrid;
-
-template <typename ValueType, typename IndexType>
-class Sellp;
-
-template <typename ValueType, typename IndexType>
-class SparsityCsr;
-
 template <typename ValueType>
 class Dense;
+
 
 /**
  * MultiVector is a matrix format which explicitly stores all values of the
@@ -99,21 +76,7 @@ class MultiVector
 #if GINKGO_ENABLE_HALF && GINKGO_ENABLE_BFLOAT16
       public ConvertibleTo<MultiVector<next_precision<ValueType, 3>>>,
 #endif
-      public ConvertibleTo<Coo<ValueType, int32>>,
-      public ConvertibleTo<Coo<ValueType, int64>>,
-      public ConvertibleTo<Csr<ValueType, int32>>,
-      public ConvertibleTo<Csr<ValueType, int64>>,
       public ConvertibleTo<Dense<ValueType>>,
-      public ConvertibleTo<Ell<ValueType, int32>>,
-      public ConvertibleTo<Ell<ValueType, int64>>,
-      public ConvertibleTo<Fbcsr<ValueType, int32>>,
-      public ConvertibleTo<Fbcsr<ValueType, int64>>,
-      public ConvertibleTo<Hybrid<ValueType, int32>>,
-      public ConvertibleTo<Hybrid<ValueType, int64>>,
-      public ConvertibleTo<Sellp<ValueType, int32>>,
-      public ConvertibleTo<Sellp<ValueType, int64>>,
-      public ConvertibleTo<SparsityCsr<ValueType, int32>>,
-      public ConvertibleTo<SparsityCsr<ValueType, int64>>,
       public DiagonalExtractable<ValueType>,
       public ReadableFromMatrixData<ValueType, int32>,
       public ReadableFromMatrixData<ValueType, int64>,
@@ -124,23 +87,8 @@ class MultiVector
       public Permutable<int64>,
       public EnableAbsoluteComputation<remove_complex<MultiVector<ValueType>>>,
       public ScaledIdentityAddable {
-    friend class EnableCloneable<MultiVector>;
-    friend class Coo<ValueType, int32>;
-    friend class Coo<ValueType, int64>;
-    friend class Csr<ValueType, int32>;
-    friend class Csr<ValueType, int64>;
     friend class Dense<ValueType>;
     friend class Diagonal<ValueType>;
-    friend class Ell<ValueType, int32>;
-    friend class Ell<ValueType, int64>;
-    friend class Fbcsr<ValueType, int32>;
-    friend class Fbcsr<ValueType, int64>;
-    friend class Hybrid<ValueType, int32>;
-    friend class Hybrid<ValueType, int64>;
-    friend class Sellp<ValueType, int32>;
-    friend class Sellp<ValueType, int64>;
-    friend class SparsityCsr<ValueType, int32>;
-    friend class SparsityCsr<ValueType, int64>;
     friend class MultiVector<to_complex<ValueType>>;
     friend class EnableCloneable<MultiVector>;
     friend class experimental::distributed::Vector<ValueType>;
@@ -152,36 +100,8 @@ public:
     using EnableCloneable<MultiVector>::move_to;
     using ConvertibleTo<MultiVector<next_precision<ValueType>>>::convert_to;
     using ConvertibleTo<MultiVector<next_precision<ValueType>>>::move_to;
-    using ConvertibleTo<Coo<ValueType, int32>>::convert_to;
-    using ConvertibleTo<Coo<ValueType, int32>>::move_to;
-    using ConvertibleTo<Coo<ValueType, int64>>::convert_to;
-    using ConvertibleTo<Coo<ValueType, int64>>::move_to;
-    using ConvertibleTo<Csr<ValueType, int32>>::convert_to;
-    using ConvertibleTo<Csr<ValueType, int32>>::move_to;
-    using ConvertibleTo<Csr<ValueType, int64>>::convert_to;
-    using ConvertibleTo<Csr<ValueType, int64>>::move_to;
     using ConvertibleTo<Dense<ValueType>>::convert_to;
     using ConvertibleTo<Dense<ValueType>>::move_to;
-    using ConvertibleTo<Ell<ValueType, int32>>::convert_to;
-    using ConvertibleTo<Ell<ValueType, int32>>::move_to;
-    using ConvertibleTo<Ell<ValueType, int64>>::convert_to;
-    using ConvertibleTo<Ell<ValueType, int64>>::move_to;
-    using ConvertibleTo<Fbcsr<ValueType, int32>>::convert_to;
-    using ConvertibleTo<Fbcsr<ValueType, int32>>::move_to;
-    using ConvertibleTo<Fbcsr<ValueType, int64>>::convert_to;
-    using ConvertibleTo<Fbcsr<ValueType, int64>>::move_to;
-    using ConvertibleTo<Hybrid<ValueType, int32>>::convert_to;
-    using ConvertibleTo<Hybrid<ValueType, int32>>::move_to;
-    using ConvertibleTo<Hybrid<ValueType, int64>>::convert_to;
-    using ConvertibleTo<Hybrid<ValueType, int64>>::move_to;
-    using ConvertibleTo<Sellp<ValueType, int32>>::convert_to;
-    using ConvertibleTo<Sellp<ValueType, int32>>::move_to;
-    using ConvertibleTo<Sellp<ValueType, int64>>::convert_to;
-    using ConvertibleTo<Sellp<ValueType, int64>>::move_to;
-    using ConvertibleTo<SparsityCsr<ValueType, int32>>::convert_to;
-    using ConvertibleTo<SparsityCsr<ValueType, int32>>::move_to;
-    using ConvertibleTo<SparsityCsr<ValueType, int64>>::convert_to;
-    using ConvertibleTo<SparsityCsr<ValueType, int64>>::move_to;
     using ReadableFromMatrixData<ValueType, int32>::read;
     using ReadableFromMatrixData<ValueType, int64>::read;
 
@@ -327,65 +247,9 @@ public:
     void move_to(MultiVector<next_precision<ValueType, 3>>* result) override;
 #endif
 
-    void convert_to(Coo<ValueType, int32>* result) const override;
-
-    void move_to(Coo<ValueType, int32>* result) override;
-
-    void convert_to(Coo<ValueType, int64>* result) const override;
-
-    void move_to(Coo<ValueType, int64>* result) override;
-
-    void convert_to(Csr<ValueType, int32>* result) const override;
-
-    void move_to(Csr<ValueType, int32>* result) override;
-
-    void convert_to(Csr<ValueType, int64>* result) const override;
-
-    void move_to(Csr<ValueType, int64>* result) override;
-
     void convert_to(Dense<ValueType>* result) const override;
 
     void move_to(Dense<ValueType>* result) override;
-
-    void convert_to(Ell<ValueType, int32>* result) const override;
-
-    void move_to(Ell<ValueType, int32>* result) override;
-
-    void convert_to(Ell<ValueType, int64>* result) const override;
-
-    void move_to(Ell<ValueType, int64>* result) override;
-
-    void convert_to(Fbcsr<ValueType, int32>* result) const override;
-
-    void move_to(Fbcsr<ValueType, int32>* result) override;
-
-    void convert_to(Fbcsr<ValueType, int64>* result) const override;
-
-    void move_to(Fbcsr<ValueType, int64>* result) override;
-
-    void convert_to(Hybrid<ValueType, int32>* result) const override;
-
-    void move_to(Hybrid<ValueType, int32>* result) override;
-
-    void convert_to(Hybrid<ValueType, int64>* result) const override;
-
-    void move_to(Hybrid<ValueType, int64>* result) override;
-
-    void convert_to(Sellp<ValueType, int32>* result) const override;
-
-    void move_to(Sellp<ValueType, int32>* result) override;
-
-    void convert_to(Sellp<ValueType, int64>* result) const override;
-
-    void move_to(Sellp<ValueType, int64>* result) override;
-
-    void convert_to(SparsityCsr<ValueType, int32>* result) const override;
-
-    void move_to(SparsityCsr<ValueType, int32>* result) override;
-
-    void convert_to(SparsityCsr<ValueType, int64>* result) const override;
-
-    void move_to(SparsityCsr<ValueType, int64>* result) override;
 
     void read(const mat_data64& data) override;
 
@@ -1375,27 +1239,6 @@ protected:
                                        this->get_const_values()),
             this->get_stride());
     }
-
-    template <typename IndexType>
-    void convert_impl(Coo<ValueType, IndexType>* result) const;
-
-    template <typename IndexType>
-    void convert_impl(Csr<ValueType, IndexType>* result) const;
-
-    template <typename IndexType>
-    void convert_impl(Ell<ValueType, IndexType>* result) const;
-
-    template <typename IndexType>
-    void convert_impl(Fbcsr<ValueType, IndexType>* result) const;
-
-    template <typename IndexType>
-    void convert_impl(Hybrid<ValueType, IndexType>* result) const;
-
-    template <typename IndexType>
-    void convert_impl(Sellp<ValueType, IndexType>* result) const;
-
-    template <typename IndexType>
-    void convert_impl(SparsityCsr<ValueType, IndexType>* result) const;
 
     /**
      * @copydoc scale(const LinOp *)
