+++ include/ginkgo/core/matrix/sellp.hpp
@@ -20,7 +20,7 @@ constexpr int default_stride_factor = 1;
 
 
 template <typename ValueType>
-class MultiVector;
+class Dense;
 
 template <typename ValueType, typename IndexType>
 class Csr;
@@ -51,7 +51,7 @@ class Sellp
 #if GINKGO_ENABLE_HALF && GINKGO_ENABLE_BFLOAT16
       public ConvertibleTo<Sellp<next_precision<ValueType, 3>, IndexType>>,
 #endif
-      public ConvertibleTo<MultiVector<ValueType>>,
+      public ConvertibleTo<Dense<ValueType>>,
       public ConvertibleTo<Csr<ValueType, IndexType>>,
       public DiagonalExtractable<ValueType>,
       public ReadableFromMatrixData<ValueType, IndexType>,
@@ -59,7 +59,7 @@ class Sellp
       public EnableAbsoluteComputation<
           remove_complex<Sellp<ValueType, IndexType>>> {
     friend class EnableCloneable<Sellp>;
-    friend class MultiVector<ValueType>;
+    friend class Dense<ValueType>;
     friend class Csr<ValueType, IndexType>;
     friend class Sellp<to_complex<ValueType>, IndexType>;
     GKO_ASSERT_SUPPORTED_VALUE_AND_INDEX_TYPE;
@@ -70,8 +70,8 @@ public:
     using ConvertibleTo<
         Sellp<next_precision<ValueType>, IndexType>>::convert_to;
     using ConvertibleTo<Sellp<next_precision<ValueType>, IndexType>>::move_to;
-    using ConvertibleTo<MultiVector<ValueType>>::convert_to;
-    using ConvertibleTo<MultiVector<ValueType>>::move_to;
+    using ConvertibleTo<Dense<ValueType>>::convert_to;
+    using ConvertibleTo<Dense<ValueType>>::move_to;
     using ConvertibleTo<Csr<ValueType, IndexType>>::convert_to;
     using ConvertibleTo<Csr<ValueType, IndexType>>::move_to;
     using ReadableFromMatrixData<ValueType, IndexType>::read;
@@ -120,9 +120,9 @@ public:
         Sellp<next_precision<ValueType, 3>, IndexType>* result) override;
 #endif
 
-    void convert_to(MultiVector<ValueType>* other) const override;
+    void convert_to(Dense<ValueType>* other) const override;
 
-    void move_to(MultiVector<ValueType>* other) override;
+    void move_to(Dense<ValueType>* other) override;
 
     void convert_to(Csr<ValueType, IndexType>* other) const override;
