+++ include/ginkgo/core/preconditioner/jacobi.hpp
@@ -186,15 +186,15 @@ struct block_interleaved_storage_scheme {
 template <typename ValueType = default_precision, typename IndexType = int32>
 class Jacobi : public LinOp,
                public EnableCloneable<Jacobi<ValueType, IndexType>>,
-               public ConvertibleTo<matrix::MultiVector<ValueType>>,
+               public ConvertibleTo<matrix::Dense<ValueType>>,
                public WritableToMatrixData<ValueType, IndexType>,
                public Transposable {
     friend class EnableCloneable<Jacobi<ValueType, IndexType>>;
     GKO_ASSERT_SUPPORTED_VALUE_AND_INDEX_TYPE;
 
 public:
-    using ConvertibleTo<matrix::MultiVector<ValueType>>::convert_to;
-    using ConvertibleTo<matrix::MultiVector<ValueType>>::move_to;
+    using ConvertibleTo<matrix::Dense<ValueType>>::convert_to;
+    using ConvertibleTo<matrix::Dense<ValueType>>::move_to;
     using value_type = ValueType;
     using index_type = IndexType;
     using mat_data = matrix_data<ValueType, IndexType>;
@@ -264,9 +264,9 @@ public:
         return blocks_.get_size();
     }
 
-    void convert_to(matrix::MultiVector<value_type>* result) const override;
+    void convert_to(matrix::Dense<value_type>* result) const override;
 
-    void move_to(matrix::MultiVector<value_type>* result) override;
+    void move_to(matrix::Dense<value_type>* result) override;
 
     void write(mat_data& data) const override;
