+++ omp/matrix/multivector_kernels.cpp
@@ -10,16 +10,6 @@
 
 #include <ginkgo/core/base/array.hpp>
 #include <ginkgo/core/base/math.hpp>
-#include <ginkgo/core/base/range_accessors.hpp>
-#include <ginkgo/core/matrix/csr.hpp>
-#include <ginkgo/core/matrix/diagonal.hpp>
-#include <ginkgo/core/matrix/fbcsr.hpp>
-#include <ginkgo/core/matrix/sellp.hpp>
-#include <ginkgo/core/matrix/sparsity_csr.hpp>
-
-#include "accessor/block_col_major.hpp"
-#include "accessor/range.hpp"
-#include "core/components/prefix_sum_kernels.hpp"
 
 
 namespace gko {
@@ -75,269 +65,6 @@ GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
     GKO_DECLARE_MULTIVECTOR_COMPUTE_NORM2_DISPATCH_KERNEL);
 
 
-template <typename ValueType, typename IndexType>
-void convert_to_coo(std::shared_ptr<const DefaultExecutor> exec,
-                    matrix::view::dense<const ValueType> source,
-                    const int64* row_ptrs,
-                    matrix::view::coo<ValueType, IndexType> result)
-{
-    auto num_rows = result.size[0];
-    auto num_cols = result.size[1];
-    auto num_nonzeros = result.num_stored_elements;
-
-    auto row_idxs = result.row_idxs;
-    auto col_idxs = result.col_idxs;
-    auto values = result.values;
-
-#pragma omp parallel for
-    for (size_type row = 0; row < num_rows; ++row) {
-        auto idxs = row_ptrs[row];
-        for (size_type col = 0; col < num_cols; ++col) {
-            auto val = source(row, col);
-            if (is_nonzero(val)) {
-                row_idxs[idxs] = row;
-                col_idxs[idxs] = col;
-                values[idxs] = val;
-                ++idxs;
-            }
-        }
-    }
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_CONVERT_TO_COO_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void convert_to_csr(std::shared_ptr<const DefaultExecutor> exec,
-                    matrix::view::dense<const ValueType> source,
-                    matrix::view::csr<ValueType, IndexType> result)
-{
-    auto num_rows = result.size[0];
-    auto num_cols = result.size[1];
-    auto num_nonzeros = result.num_stored_elements;
-
-    auto row_ptrs = result.row_ptrs;
-    auto col_idxs = result.col_idxs;
-    auto values = result.values;
-
-#pragma omp parallel for
-    for (size_type row = 0; row < num_rows; ++row) {
-        auto cur_ptr = row_ptrs[row];
-        for (size_type col = 0; col < num_cols; ++col) {
-            auto val = source(row, col);
-            if (is_nonzero(val)) {
-                col_idxs[cur_ptr] = col;
-                values[cur_ptr] = val;
-                ++cur_ptr;
-            }
-        }
-    }
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_CONVERT_TO_CSR_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void convert_to_ell(std::shared_ptr<const DefaultExecutor> exec,
-                    matrix::view::dense<const ValueType> source,
-                    matrix::view::ell<ValueType, IndexType> result)
-{
-    auto num_rows = result.size[0];
-    auto num_cols = result.size[1];
-    auto max_nnz_per_row = result.num_stored_elements_per_row;
-#pragma omp parallel for
-    for (size_type i = 0; i < max_nnz_per_row; i++) {
-        for (size_type j = 0; j < num_rows; j++) {
-            result.val_at(j, i) = zero<ValueType>();
-            result.col_at(j, i) = invalid_index<IndexType>();
-        }
-    }
-#pragma omp parallel for
-    for (size_type row = 0; row < num_rows; row++) {
-        size_type col_idx = 0;
-        for (size_type col = 0; col < num_cols; col++) {
-            auto val = source(row, col);
-            if (is_nonzero(val)) {
-                result.val_at(row, col_idx) = val;
-                result.col_at(row, col_idx) = col;
-                col_idx++;
-            }
-        }
-    }
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_CONVERT_TO_ELL_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void convert_to_fbcsr(std::shared_ptr<const DefaultExecutor> exec,
-                      matrix::view::dense<const ValueType> source,
-                      matrix::Fbcsr<ValueType, IndexType>* result)
-{
-    const auto num_rows = source.size[0];
-    const auto num_cols = source.size[1];
-    const auto bs = result->get_block_size();
-    const auto nzbs = result->get_num_stored_blocks();
-    const auto num_block_rows = num_rows / bs;
-    const auto num_block_cols = num_cols / bs;
-    acc::range<acc::block_col_major<ValueType, 3>> blocks(
-        std::array<acc::size_type, 3>{static_cast<acc::size_type>(nzbs),
-                                      static_cast<acc::size_type>(bs),
-                                      static_cast<acc::size_type>(bs)},
-        result->get_values());
-    auto col_idxs = result->get_col_idxs();
-#pragma omp parallel for
-    for (size_type brow = 0; brow < num_block_rows; ++brow) {
-        auto block = result->get_const_row_ptrs()[brow];
-        for (size_type bcol = 0; bcol < num_block_cols; ++bcol) {
-            bool block_nz = false;
-            for (int lrow = 0; lrow < bs; ++lrow) {
-                for (int lcol = 0; lcol < bs; ++lcol) {
-                    const auto row = lrow + bs * brow;
-                    const auto col = lcol + bs * bcol;
-                    block_nz = block_nz || is_nonzero(source(row, col));
-                }
-            }
-            if (block_nz) {
-                col_idxs[block] = bcol;
-                for (int lrow = 0; lrow < bs; ++lrow) {
-                    for (int lcol = 0; lcol < bs; ++lcol) {
-                        const auto row = lrow + bs * brow;
-                        const auto col = lcol + bs * bcol;
-                        blocks(block, lrow, lcol) = source(row, col);
-                    }
-                }
-                block++;
-            }
-        }
-    }
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_CONVERT_TO_FBCSR_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void convert_to_hybrid(std::shared_ptr<const DefaultExecutor> exec,
-                       matrix::view::dense<const ValueType> source,
-                       const int64* coo_row_ptrs,
-                       matrix::view::hybrid<ValueType, IndexType> result)
-{
-    auto num_rows = result.size[0];
-    auto num_cols = result.size[1];
-    auto ell_lim = result.ell_part.num_stored_elements_per_row;
-    auto coo_val = result.coo_part.values;
-    auto coo_col = result.coo_part.col_idxs;
-    auto coo_row = result.coo_part.row_idxs;
-
-#pragma omp parallel for
-    for (size_type row = 0; row < num_rows; row++) {
-        size_type ell_count = 0;
-        size_type col = 0;
-        for (; col < num_cols && ell_count < ell_lim; col++) {
-            auto val = source(row, col);
-            if (is_nonzero(val)) {
-                result.ell_part.val_at(row, ell_count) = val;
-                result.ell_part.col_at(row, ell_count) = col;
-                ell_count++;
-            }
-        }
-        for (; ell_count < ell_lim; ell_count++) {
-            result.ell_part.val_at(row, ell_count) = zero<ValueType>();
-            result.ell_part.col_at(row, ell_count) = invalid_index<IndexType>();
-        }
-        auto coo_idx = coo_row_ptrs[row];
-        for (; col < num_cols; col++) {
-            auto val = source(row, col);
-            if (is_nonzero(val)) {
-                coo_val[coo_idx] = val;
-                coo_col[coo_idx] = col;
-                coo_row[coo_idx] = row;
-                coo_idx++;
-            }
-        }
-    }
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_CONVERT_TO_HYBRID_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void convert_to_sellp(std::shared_ptr<const DefaultExecutor> exec,
-                      matrix::view::dense<const ValueType> source,
-                      matrix::view::sellp<ValueType, IndexType> result)
-{
-    const auto num_rows = result.size[0];
-    const auto num_cols = result.size[1];
-    const auto vals = result.values;
-    const auto col_idxs = result.col_idxs;
-    const auto slice_sets = result.slice_sets;
-    const auto slice_size = result.slice_size;
-    const auto num_slices = ceildiv(num_rows, slice_size);
-#pragma omp parallel for
-    for (size_type slice = 0; slice < num_slices; slice++) {
-        for (size_type local_row = 0; local_row < slice_size; local_row++) {
-            const auto row = slice * slice_size + local_row;
-            if (row >= num_rows) {
-                break;
-            }
-            auto sellp_idx = slice_sets[slice] * slice_size + local_row;
-            const auto sellp_end =
-                slice_sets[slice + 1] * slice_size + local_row;
-            for (size_type col = 0; col < num_cols; col++) {
-                auto val = source(row, col);
-                if (is_nonzero(val)) {
-                    col_idxs[sellp_idx] = col;
-                    vals[sellp_idx] = val;
-                    sellp_idx += slice_size;
-                }
-            }
-            for (; sellp_idx < sellp_end; sellp_idx += slice_size) {
-                col_idxs[sellp_idx] = invalid_index<IndexType>();
-                vals[sellp_idx] = zero<ValueType>();
-            }
-        }
-    }
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_CONVERT_TO_SELLP_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void convert_to_sparsity_csr(std::shared_ptr<const DefaultExecutor> exec,
-                             matrix::view::dense<const ValueType> source,
-                             matrix::SparsityCsr<ValueType, IndexType>* result)
-{
-    auto num_rows = result->get_size()[0];
-    auto num_cols = result->get_size()[1];
-
-    auto row_ptrs = result->get_row_ptrs();
-    auto col_idxs = result->get_col_idxs();
-    auto value = result->get_value();
-    value[0] = one<ValueType>();
-
-#pragma omp parallel for
-    for (size_type row = 0; row < num_rows; ++row) {
-        auto cur_ptr = row_ptrs[row];
-        for (size_type col = 0; col < num_cols; ++col) {
-            auto val = source(row, col);
-            if (is_nonzero(val)) {
-                col_idxs[cur_ptr] = col;
-                ++cur_ptr;
-            }
-        }
-    }
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_CONVERT_TO_SPARSITY_CSR_KERNEL);
-
-
 template <typename ValueType>
 void transpose(std::shared_ptr<const DefaultExecutor> exec,
                matrix::view::dense<const ValueType> orig,
@@ -369,37 +96,6 @@ void conj_transpose(std::shared_ptr<const DefaultExecutor> exec,
 
 GKO_INSTANTIATE_FOR_EACH_VALUE_TYPE(
     GKO_DECLARE_MULTIVECTOR_CONJ_TRANSPOSE_KERNEL);
-
-
-template <typename ValueType, typename IndexType>
-void count_nonzero_blocks_per_row(std::shared_ptr<const DefaultExecutor> exec,
-                                  matrix::view::dense<const ValueType> source,
-                                  int bs, IndexType* result)
-{
-    const auto num_rows = source.size[0];
-    const auto num_cols = source.size[1];
-    const auto num_block_rows = num_rows / bs;
-    const auto num_block_cols = num_cols / bs;
-#pragma omp parallel for
-    for (size_type brow = 0; brow < num_block_rows; ++brow) {
-        IndexType num_nonzero_blocks{};
-        for (size_type bcol = 0; bcol < num_block_cols; ++bcol) {
-            bool block_nz = false;
-            for (int lrow = 0; lrow < bs; ++lrow) {
-                for (int lcol = 0; lcol < bs; ++lcol) {
-                    const auto row = lrow + bs * brow;
-                    const auto col = lcol + bs * bcol;
-                    block_nz = block_nz || is_nonzero(source(row, col));
-                }
-            }
-            num_nonzero_blocks += block_nz ? 1 : 0;
-        }
-        result[brow] = num_nonzero_blocks;
-    }
-}
-
-GKO_INSTANTIATE_FOR_EACH_VALUE_AND_INDEX_TYPE(
-    GKO_DECLARE_MULTIVECTOR_COUNT_NONZERO_BLOCKS_PER_ROW_KERNEL);
 
 
 }  // namespace multivector
